requirements.txt
annotated-doc==0.0.4
annotated-types==0.6.0
anyio==4.2.0
APScheduler==3.6.3
asttokens==2.4.1
attrs==24.3.0
bcrypt==5.0.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==4.2.2
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
comtypes==1.3.1
debugpy==1.8.9
decorator==5.1.1
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
executing==2.1.0
fastapi==0.128.0
ffmpeg-python==0.2.0
filelock==3.16.1
Flask==3.0.3
fsspec==2024.10.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.3.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.26.5
idna==3.6
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.29.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.2
openai==1.14.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.4.0
prompt_toolkit==3.0.48
psutil==6.1.0
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyasn1==0.6.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.18.0
pyjokes==0.6.0
pymongo==4.13.1
pypiwin32==223
PySocks==1.7.1
pyTelegramBotAPI==4.16.1
pytest==8.0.0
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.22
python-telegram-bot==20.3
pyttsx3==2.90
pytube==15.0.0
pytz==2025.2
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.0.3
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.27.1
setuptools==69.1.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.10.1
spotipy==2.23.0
SQLAlchemy==2.0.46
sqlmodel==0.0.31
stack-data==0.6.3
starlette==0.50.0
sympy==1.13.1
telebot==0.0.5
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
transformers==4.47.0
trio==0.27.0
trio-websocket==0.11.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.2.3
uvicorn==0.40.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.3
windows-curses==2.4.1
wsproto==1.2.0
youtube-dl==2021.12.17
